more forgiving C-style line breaks (WIP)
I kept this changeset compatible with base Wren and its test suite to be extra safe. This hack allows curly braces on newlines (aka Allman style), as well as multiline conditional statements (with support for comments). It also allows line breaks without the need for curly braces at all for single-statement if/else/for/while, with support for nesting.